Update README.md
This file explains the project Simulating the Emergence of Innovation in detail. It is found that when ensuring that ideas can be generated, decreasing the probability of idea generation and increasing the probability of social interaction can help to foster innovation. The finding suggests that creating a culture so that people are comfortable sharing ideas is helpful for society. Also, educating people so that ideas are generated focusing on quality, not quantity, can foster innovation.